longAddedinRepo

+++ src/main/java/com/alkemy/ong/data/repository/ActivityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActivityRepository extends JpaRepository<ActivityEntity, Integer> {
+public interface ActivityRepository extends JpaRepository<ActivityEntity, Long> {
 }
